mel: enable copyleft for native/cross/etc
This is technically only needed for our release builds, but we may as well be
proactive and inclusive about it. Without this, copyleft_compliance was
defaulting to just target, and as a result sources like autoconf, automake
were missing from the archived DL_DIR that ends up in our installers.

Signed-off-by: Christopher Larson <chris_larson@mentor.com>