Add support for USER_FEATURES
USER_FEATURES is a mechanism by which a user can exert control over
DISTRO_FEATURES from local.conf.

To add a feature, simply add it to USER_FEATURES. To remove a feature, add the
feature prefixed by ~.

Example:

    # I don't want bluetooth support, but I want to make sure I build things
    # with x11 support
    USER_FEATURES += "~bluetooth x11"

There is one side effect of this: it forces early expansion of DISTRO_FEATURES
at ConfigParsed time, as it has to expand it to get the list to filter.

Signed-off-by: Christopher Larson <chris_larson@mentor.com>