mel: drop the largefile bits for now
We can't use oe_filter_out here due to parse ordering issues. We can't use
USER_FEATURES, because the poky DISTRO_FEATURES_append happens after that
processing. So drop this for now, and if we need to bring back this handling,
we can get it directly into meta-yocto/conf/poky.conf.

Signed-off-by: Christopher Larson <chris_larson@mentor.com>