netbase: drop interfaces file
It's not needed, and it's causing problems with the qemu machines.

When filespath is constructed, it's assembled based on the layer path first,
then the override, which means that less specific override files in higher
priority layers win out over more specific override files in lower priority
layers, at this time. I'll be investigating a change to that behavior for
upstream as well, if possible.

Signed-off-by: Christopher Larson <chris_larson@mentor.com>