autogen-native: fix possible ubuntu contamination
The m4 macro which determines the guile version initially checks dpkg --list
before running guile-config --version (ugh), so if your debian or ubuntu
system has guile-1.8-libs installed, it will assume 1.8 when we have 2.0 in
the sysroots, and may fail to compile. Fix by getting the version ourselves
and passing it in as a cached test result.

Signed-off-by: Christopher Larson <chris_larson@mentor.com>