liburcu: upgrade from 0.6.7 to 0.7.5
0.7.4 added support for MIPS, so this reduces the delta of us vs upstream.
It's also required for the git versions of the lttng 2 recipes.

Note that the license checksum changed because of a few minor changes (to the
lists of individual files in LICENSE) and because of an addition of a GPLv3
macro. As this macro isn't shipped in any form, I've not included that in
LICENSE. In addition, even though there are also GPLv2 files in the tree (the
library tests), we don't ship those, so that's left out as well. And finally,
though there are MIT files in the source tree, they're linked against LGPL
sources, so that license is superceded.

Signed-off-by: Christopher Larson <chris_larson@mentor.com>