setup-builddir: add full support for LAYERDEPENDS
There was substantial refactoring to pull this off. We switched from sh to
bash so we can leverage associative arrays to store layer information we
extract from layer.conf.

If a required layer's dependency exists, we include it.
If a required layer's dependency does not exist, we notify and abort.
We only include an optional layer if all of its dependencies are already
included.

Signed-off-by: Christopher Larson <chris_larson@mentor.com>