setup-builddir: add full support for LAYERDEPENDS
There was substantial refactoring to pull this off. We switched from sh to
bash so we can leverage associative arrays to store layer information we
extract from layer.conf.

If a required layer's dependency exists, we include it.
If a required layer's dependency does not exist, we notify and abort.
We only include an optional layer if all of its dependencies are already
included.

The one difference between master's setup-builddir and mel6's setup-builddir
is the latter currently doesn't lower meta-oe's priority below oe-core, as
doing so could have an impact/risk given this is supposed to be a bugfix
branch.

Signed-off-by: Christopher Larson <chris_larson@mentor.com>