archive-release: make mel-release a bit more generic
The only remaining references to MEL are a compatibility entry in PROVIDES,
and the use of MELDIR, which is currently necessary as we need to know the
root path to the layers for the layer archival process. This could be renamed
to OEROOT or OEDIR or similar in the long term, but this will do for now.

Also add default values for RELEASE_EXCLUDED_SOURCES and COPYLEFT_SOURCES_DIR,
and add a number of 'doc' flags for the variables used by the recipe.

JIRA: SB-913

Signed-off-by: Christopher Larson <chris_larson@mentor.com>