setup-environment: always process bblayers.conf
This ensures that e.g. a reconfiguration of an existing build dir for the same
machine (no machine specified) still results in the correct set of layers
given current OPTIONALLAYERS. It also ensures that the differences between the
mel and atp setup scripts are retained even for the customer with both
installed.

Signed-off-by: Christopher Larson <chris_larson@mentor.com>