gpu-viv-bin-mx6q: Rework package for a single backend use
As we now provide a single backend there're no need for full set of
packages and the binaries can use the generic names so we don't need
to rework other applications and/or libraries to link properly.

The xf86-video-imxfb-vivante needs also to be change as the libraries
are now under generic name so we remove the backend suffix from
libraries so the LDFLAGS ought to be removed for the build to succeed.

Change-Id: I73cc9986babe8bb4fb0c05117c6b3d93b0470949
Signed-off-by: Otavio Salvador <otavio@ossystems.com.br>

Pulled from meta-fsl-arm as a bbappend. -CL

Signed-off-by: Christopher Larson <chris_larson@mentor.com>