Work around sstate_create_package / do_rootfs race
do_rootfs adds ${S} to its cleandirs (unnecessarily), and
sstate_create_package doesn't have a 'dirs' flag set. Bitbake apparently
defaults to a dirs of ${B} when dirs is unset (bad, very bad), so it's
possible for sstate_create_package to try to cd into a dir which is being
removed.

Signed-off-by: Christopher Larson <chris_larson@mentor.com>