mel: run do_populate_lic for deps of an image
Ensure that the license manifest is complete. Currently, the manifest code
skips emission of licenses which aren't in deploy/licenses, and does so
nearly silently. A warning is printed, but it's only visible in the
do_rootfs task log. This ensures the populate_lic task is run for the deps
of an image, as a workaround.

I expect that the correct fix upstream would be to alter the manifest
generation behavior for the case where the generic license files aren't in
deploy/licenses.

JIRA: SB-1122

Signed-off-by: Christopher Larson <chris_larson@mentor.com>