package.bbclass: make exclusion warning a note instead
This isn't a case of something questionable occurring, it's a case of the
configuration being obeyed, which is as it should be. Make it a note. The user
can still examine the do_package log if they want to know why a package isn't
being emitted.

JIRA: SB-1144

Signed-off-by: Christopher Larson <chris_larson@mentor.com>
(cherry picked from commit 914cf0642cea6b0784f3d9fbca6d1930dadccc34)

Signed-off-by: Michael Brown <mw_brown@mentor.com>