sanity: fix incorrect 'without march' logic in check_gcc_march
The 'result' variable was used both for the output of the gcc command run, and
for the boolean value being returned. As a result, when the first gcc command
succeeded, it returned that output, so the caller would interpret the result
as False when the gcc command output was the empty string, which is obviously
not the intention.

Signed-off-by: Christopher Larson <chris_larson@mentor.com>