cml1-config: add class for .config setup
It sets up the .config from defconfig in a consistent way, and uses
kern-tools-native to merge in config fragments the way busybox and linux-yocto
do. This should ideally go upstream. We'll want to adapt our non-linux-yocto
kernels in the BSPs we support so we can utilize fragments, and so we can
enable the lttng support which we require in a clean way.

Signed-off-by: Christopher Larson <kergoth@gmail.com>