Use cml1-config to configure kernels for lttng
The new method kills the global kernel-config-fragments inherit, as that had
particular expectations about use of a defconfig variable, instead adding
a new kernel-lttng bbclass which uses cml1-config to merge in the lttng
configuration. This does mean we'd need to bbappend a bit more to enable it in
particular kernels, but this makes it more explicit, and long term should be
unnecessary if we can get cml1-config upstream and alter bsp layer kernels to
use it.

Signed-off-by: Christopher Larson <kergoth@gmail.com>