qtwebkit: export RUBYLIB so our do_compile behaves
do_compile runs ruby from ruby-native, but it seems that ruby either never had
the correct search paths into the sysroot, or it has hardcoded paths and
therefore has a relocation bug. Exporting an explicit RUBYLIB to hit the
appropriate sysroot paths works around the problem. This workaround was copied
from the mkvtoolnix recipe in meta-oe.

Signed-off-by: Christopher Larson <chris_larson@mentor.com>