Added handling for customer repos, hotfixes, etc.
Added three sections to parse the following options config files
for additional layers to add to the config:

	mel/customer.conf
	mel/xlayers.conf
	mel/hotfixes/hotfix.conf

All three are optional.  customer.conf must contain the customer name
and will pull in the custom configuration for "adit-custom" repo, for
example when "adit" is listed in customer.conf

hotfixes/hotfix.conf lists the layer names for all desired hotfixes
to be included in the build.  It does not automatically add all hotfixes
in order to pull out problematic hotfixes, and more importantly to allow
cooperative work on hotfixes.  That is, incomplete hotfixes can be merged
in without worry of mangling the nightly build until they are ready and
added to hotfix.conf.

xlayers.conf is intended to be used as an easy way for a developer to
add a particular layer to the build configuration without having to
go and modify OPTIONALLAYERS inside setup-environment.  This can't be
done within local.conf because bblayers.conf is created before local.conf
is ever created/parsed.