bb-determine-layers: make the layer order deterministic
Previously, we sorted by priority only. As a result, layers that shared the
same priority would be ordered in a non-deterministic way -- whatever order we
found the layer files on disk in the glob expansion.

Because of this, if we have layers with the same priority that append to the
same recipe, the append order would be non-deterministic, and we could end up
not successfully reusing sstates due to e.g. changing order of items in
SRC_URI.

Alter it so layers with the same priority will be sorted by name.

Signed-off-by: Christopher Larson <kergoth@gmail.com>