base.bbclass: fix nondeterministic pkgconfig order
The PACKAGECONFIG flags were iterated over using dict.items(), but this
returns the items in an undefined order. As this order determines the
EXTRA_OECONF append order, we can get EXTRA_OECONF which are functionally
equivalent, but whose contents differ, resulting in not using shared state
archives we should be using.

JIRA: MEIP-165, SB-1462

Signed-off-by: Christopher Larson <kergoth@gmail.com>