opkg-utils: allow disabling update-alternatives
This is needed to deal with the situation where we're using ipk packaging, so
opkg-utils must be built regardless of what update-alternatives provider we
prefer. The downside to the current implementation is the need to adjust
PACKAGECONFIG as well as PREFERRED_PROVIDER, but it is more explicit that way.

JIRA: SB-1902

Signed-off-by: Christopher Larson <kergoth@gmail.com>