archive-release-downloads: switch to this
This and copyleft_compliance essentially do the same thing, but the purposes
are different (here we want to populate a DL_DIR, there is about copyleft
license compliance), and we want to let the user use copyleft_compliance the
way they choose to, rather than repurposing it for our releases.

This class is also opt-in, in that the task isn't run by default even when
inherited, unless you're building archive-release.

Signed-off-by: Christopher Larson <kergoth@gmail.com>