mel: add STAMPCLEAN to BB_HASHBASE_WHITELIST
This keeps STAMPS_DIR from leaking into the do_configure checksum, which
disrupts the ability to use bitbake -S or bitbake-whatchanged to examine
checksum differences, at a minimum. This is going upstream, courtesy Robert
Yang, but let's work around it in the meantime. Drop when we update our
upstream layers again.

JIRA: SB-2044

Signed-off-by: Christopher Larson <kergoth@gmail.com>