glib-2.0: Do not use readlink when cross compiling.
When configuring glib, the configure.ac script uses readlink
to convert a relative path to an absolute path.  In my particular
build system, /usr/lib64 is a symlink to /usr/lib so the files
get packaged in the wrong location resulting in the following:

    WARNING: QA Issue: glib-2.0-dbg: found library in wrong location: /usr/share/gdb/auto-load/usr/lib/libgobject-2.0.so.0.3800.2-gdb.py
    glib-2.0-dbg: found library in wrong location: /usr/share/gdb/auto-load/usr/lib/libglib-2.0.so.0.3800.2-gdb.py

This patches configure.ac to simply use the path unmodified.
In the default configuration it is not a relative path so
there is no issue with it needing to be converted.

Submitted to the Gnome developers here:
    https://bugzilla.gnome.org/show_bug.cgi?id=727575

Signed-off-by: Drew Moseley <drew_moseley@mentor.com>