trace-cmd: fix target paths
It was incorrectly picking up our absolute 'libdir', but in the trace-cmd
makefile, 'libdir' is just the relative path from prefix, 'lib'. As a result,
we ended up with plugins in /usr/usr/lib/trace-cmd/plugins. Pass that in as
a correct relative path, and while we're at it, ensure it obeys our paths for
mandir, bindir, etc.

JIRA: SB-2580

Signed-off-by: Christopher Larson <kergoth@gmail.com>