mel: don't inherit isolated-sstate-dir by default
Now that the default SSTATE_DIR is under ${TOPDIR}, pulling from MELDIR via
SSTATE_MIRRORS, this is rather unnecessary, as we'll already have an sstate
directory isolated to this build directory, if not isolated to TMPDIR. We can
keep isolated-sstate-dir.bbclass around, as it's still useful in certain
cases. For example, my internal configuration is to write to
MELDIR/cached-binaries, not to TOPDIR/cached-binaries, so my cache is shared,
not isolated.

JIRA: SB-2614

Signed-off-by: Christopher Larson <kergoth@gmail.com>