mel: fix IMAGE_FEATURES vardeps bug
Changes to IMAGE_FEATURES or FEATURE_PACKAGE_<feature> for each feature in
IMAGE_FEATURES didn't result in a change to the do_rootfs checksum, so the
task wouldn't re-run when those variables changed. Fix by expanding the inline
python as the vardepvalue, so any change to the package list to be installed
based on image features will affect the checksum.

JIRA: SB-3056

Signed-off-by: Christopher Larson <kergoth@gmail.com>