kernel-module-split: drop seemingly pointless change
We had this class modified to let us change what package the kernel modules
depend upon (normally kernel-<kernel version>), but this is pointless. The
modules indirectly depend on kernel-base, which depends on kernel-image. If
the user doesn't want the kernel in the rootfs (or better yet, the bsp/machine
doesn't), then RDEPENDS_kernel-base is specifically allowed to be overridden
(see kernel.bbclass).

Signed-off-by: Christopher Larson <chris_larson@mentor.com>