kernel-host-dev: Host-built binaries from kernel/scripts/
This recipe provides the missing bits for /usr/src/kernel/scripts/ to do out
of tree cross kernel module builds. This target package is only suitable for
the ADE, not a target image, and assumes the host bitbake is run on is the
same architecture as the host the ADE is installed on.

JIRA: SB-3097

Signed-off-by: Christopher Larson <chris_larson@mentor.com>