setup-environment: Default to mel-lite
For now, set the MEL_DISTRO variable to mel-lite to enable
the appropriate config in the Jenkins build.  This will be
the only difference between this branch and the stock release/2014.05-async1
branch and is only an interim fix for the release/2014.05 branches.
We will need a more robust solution for the winter release.

Signed-off-by: Drew Moseley <drew_moseley@mentor.com>