setup-environment: add local.conf snippet support
For each configured layer, check for conf/local.conf.append, and if it exists,
append it to local.conf, in a bracketed block. It will automatically remove an
existing block for this file if it exists, to handle reconfiguration of
existing build directories. This should be much more convenient than using
post-setup-environment for the most common case.

Signed-off-by: Christopher Larson <chris_larson@mentor.com>