consolekit: use systemd_unitdir for modifications
In case of 64-bit systems the base_libdir is set to lib64 whereas
the systemd units are still installed into to lib/systemd so usage
of base_libdir for modifications to unit files breaks on 64-bit
systems.
We now use the systemd_unitdir variable for unit files path which
is the same passed for installation to the package.

JIRA Ticket: SB-4003

Signed-off-by: Awais Belal <awais_belal@mentor.com>