alsa-lib: Protect from freeing semaphore when already in use
In the case of dshare, dsnoop, and dmix when a device is opened twice
and fails the second time, the semaphore is completely discarded. This
creates dangling semaphore data.

This patch removes the possibility for the semaphore to be destroyed during
a typical open failure by first checking if the shared memory can be destroyed
or not. If the shared memory cannot be released it means both it and the
semaphore are still in use and therefore the semaphore is just released.

Port of hf2013.11-127-alsa-protect-semaphore

JIRA: BGTHREE-232

Signed-off-by: Joshua Frkuska <joshua_frkuska@mentor.com>
Signed-off-by: Mikhail Durnev <mikhail_durnev@mentor.com>