meta-mentor-staging/conf/layer.conf: search for *.bb files
There is a possibility that a new recipe is to be pushed upstream
or such modifications to existing recipe are required for which
append file can't be created and creating the modified copy is the
only option e.g. modifying the python function that gets executed
during parse time. This commit relaxes search so that *.bb also
get picked.

Signed-off-by: Fahad Arslan <Fahad_Arslan@mentor.com>