buildtools-tarball: drop no longer necessary workarounds
We no longer need the explicit perl module package inclusions, as upstream
switched to pulling in `python-modules`, which is all the modules which come
from python itself.

Also we can drop the bits to forcibly remove perl, as upstream dealt with that
when they added the git perltools to the buildtools tarball.

Signed-off-by: Christopher Larson <kergoth@gmail.com>