pulseaudio: disable autospawning by default
PulseAudio provides a feature where it can autospawn itself
if an application intends to use it but this makes the
enabling/disabling of PA indeterministic from a user's point
of view. Supposing that a user disables PA through the
init manager the PA daemon should never be launched so
we disable autospawning by default.

Signed-off-by: Awais Belal <awais_belal@mentor.com>