pulseaudio: fix autostart behavior
This commit picks up some changes from PA upstream that
fix the autostart behavior for pulseaudio. Earlier to
these, the pulseaudio server would be launched through
x11 startup scripts as well as sysv/systemd scripts
which is not a very good method and makes the behavior
indeterministic.
These patches now disable the handling through x11 startup
scripts so now we only use init manager to control the
startup of pulseaudio.

Signed-off-by: Awais Belal <awais_belal@mentor.com>