mel: don't include multilib.conf
The only reason we included it by default was for consistent library directory
naming between multilib and non-multilib builds, but we could just define
baselib the way multilib.conf does to accomplish that, and we would then avoid
the duplicate inclusion when a machine pulls it in.

That said, when we make the naming consistent, then e.g. a non-multilib rootfs
for qemux86-64 will have /lib64, but no /lib. Having just a suffixed libdir
and no /lib could be potentially confusing, and potentially non-FHS-comlpiant,
as /lib<qual> is optional, but /lib is not, so just remove that behavior
entirely.

Signed-off-by: Christopher Larson <chris_larson@mentor.com>