pulseaudio: use the upstream autospawn
Running pulseaudio as a system service is not preferred, yet we can't run it
as a systemd user service since we don't use pam, so switch to upstream
pulseaudio's pre-systemd behavior, use of the built in autospawn capability.

See the freedesktop wiki for info:

https://wiki.freedesktop.org/www/Software/PulseAudio/Documentation/User/Running/

This aligns us closer to upstream pulseaudio and upstream Yocto, and reduces
what we have to maintain.

Signed-off-by: Christopher Larson <chris_larson@mentor.com>