mel: minimize POKY_DEFAULT_ overrides
Doing so will add kernel-module-af-packet to the recommends, and
packagegroup-core-boot to the depends, but neither appear to affect the
contents of any of our images, so it's not problematic. Use _remove to
explicitly remove the ptest and wayland (recently added) features, until such
time as we want to support wayland across the board in mel.

Signed-off-by: Christopher Larson <chris_larson@mentor.com>