image_types_wic: add class to fix IMAGE_CMD_wic
The upstream function hardcodes the wks path, making foolish assumptions and
really not aligning with how things are done in the layering scheme. This
alters it, as a start, to use a variable for the path. I have a local
prototype of a version that leverages a file:// wks URI to find it in the
FILESPATH rather than hardcoding a full path, as well, but am working out some
kinks in that method before merging it. This will work in the meantime, as one
can use inline python to get the wks path by searching BBLAYERS.

Signed-off-by: Christopher Larson <chris_larson@mentor.com>