gdbserver-gplv3: fix gplv3 gdbserver handling
This needed to account for the switch to granular external recipes. In
addition, upstream wipes incompatible licenses from the image license manifest
at this time (not good), so we can't check for gdbserver being installed when
ALLOW_GPLV3_GDBSERVER isn't set. Instead of always letting gdbserver-external
build, and then controlling its installability with the variable, only emit
the package at all when the var is set instead.

Signed-off-by: Christopher Larson <chris_larson@mentor.com>