archive-release: handle RELEASE_ARTIFACTS=downloads
It turns out the do_build[recrdeptask] was unconditionally sucking in the
do_package_write_* tasks for the image regardless of how the image was pulled
in, whether DEPENDS or explicit 'depends'. Be more explicit in our deps, and
remove the recrdeptask so we can properly archive-release just the downloads
for our qemu builds.

JIRA: SB-6266

Signed-off-by: Christopher Larson <chris_larson@mentor.com>