python: ensure usage of native modules while cross compiling
During build of shared modules the interpreter is trying to
import modules from the freshly built target side directories.
This fails where the generated module code is target architecture
specific and the build fails with an illegal instruction error.
The issue can be reproduced for any machine by adding the
-march=bdver4 flag inserted in the TUNE_CC_ARGS which generates
quite target specific code.
We fix this by modifying the PYTHONPATH variable which augments
the default search path.

Signed-off-by: Awais Belal <awais_belal@mentor.com>