metadata_scm.bbclass: restore capture of stderr
We don't want the user to see errors from the git commands run by metadata_scm
on their console, so we need to capture or suppress stderr as well as stdout.
This was the case prior to the rewrite of the git hash logic, but the 2>&1 was
lost when it was reworked. Bring it back to avoid messages like this in builds
with non-git layers:

    fatal: Not a git repository (or any of the parent directories): .git

JIRA: SB-6295

Cc: Ross Burton <ross.burton@intel.com>
Signed-off-by: Christopher Larson <chris_larson@mentor.com>