recipetool.kernel: enable defconfig for kernel-yocto
It turns out kernel-yocto.bbclass will use a defconfig as a baseline
configuration if it's installed to WORKDIR, so we can remove the check which
aborts this sub-command for those recipes. The configuration will likely be
altered from the baseline based on the scc files and configuration fragments,
but we can set the initial baseline configuration.

JIRA: SB-6350

Signed-off-by: Christopher Larson <chris_larson@mentor.com>