weston: build weston with rdp support for cyclone5
Configure weston to build with its RDP backend enabled.
Add freerdp to DEPENDS, as required by the RDP backend.

RDP backend is needed in order to access the wayland display
as cyclone5 does not have a physical display port. This
enables the user to connect to the remote display over
network.

JIRA SB-6540

Signed-off-by: Abdur Rehman <abdur_rehman@mentor.com>