Cyclone5: Add script to create boot partition
The SD card for Cyclone-5 needs a custom partition with type=0xa2. It is
required by the BootROM to read the boot loader binary. Currently wic does
not have support for assigning an arbitrary value to partition id. This
script create the special partition using fdisk utility. The SD-Card for
Cyclone 5 must be created using wic image before using this script.

Signed-off-by: Fahad Usman <fahad_usman@mentor.com>