libvpx: fix host contamination issue
Fixes
ERROR: QA Issue: libvpx: /libvpx/usr/lib/libvpx.so.2.0.0 is owned by uid 1923, which is the same as the user running bitbake. This may be due to host contamination [host-user-contaminated]
ERROR: QA Issue: libvpx: /libvpx-dev/usr/lib/pkgconfig/vpx.pc is owned by uid 1923, which is the same as the user running bitbake. This may be due to host contamination [host-user-contaminated]

Signed-off-by: Awais Belal <awais_belal@mentor.com>