inetlib, gnumail: remove duplicate files in favor of symlinks
oe_jarinstall() is responsible for installing jar files and
their symlinks. These recipes make a copy of a jar file at
the same location where oe_jarinstall() wants to create a
symlink.

When oe_jarinstall() sees a file that is already present at
the same path where it wants to create a symlink, it throws a
warning that file is in the way, deletes the file and then
creates the symlink.

Remove such files before oe_jarinstall() is run to quiet the
warning.

JIRA: SB-6820

Signed-off-by: Abdur Rehman <abdur_rehman@mentor.com>