mel: exclude DATETIME_SECS from checksums
Currently, bitbake is erroring with a taskbase hash mismatch due to the ade
version string changing between the main bitbake process which checks
sstate and the worker process that runs the task.

So, exclude the datetime from the checksums to avoid this mismatch. This
means a bitbake of populate_ade won't re-run every time, but when it does
re-run, the version will still increment. This retains the properties we
need of the versioning while fixing the bitbake error.

Signed-off-by: Christopher Larson <chris_larson@mentor.com>