local.conf.append.cyclone5: prefer v2.0 of xserver-nodm-init
This makes sure that we use the same version of xserver-nodm-init
from meta-oe that we verified before.

Due to a change in layer priorities as of commit id
dd27fca60ee00c5a53cbcc5c18cffd081a689754 in meta-mentor,
the build system resolves the recipe contention in favor of
poky/meta which has an older version(v1.0) of the same recipe.

Settle this by adding a preference for v2.0.

JIRA: SB-7115

Signed-off-by: Abdur Rehman <abdur_rehman@mentor.com>