t4240rdb-64b: drop BUILD_64BIT_KERNEL
This is only needed for the non-64b machine variants, as those default to 32
bit. Since this defaults to 64 bit, the kernel ends up 64 bit anyway, so
there's no need for this. Also, e6500.inc sets it by default for the 32 bit
variants, so we don't need to set it for those either.

Signed-off-by: Christopher Larson <chris_larson@mentor.com>