mel: enable wic fstypes if a wks exists
As we prefer wic images for our BSPs, we can automatically enable such images
when a wks file exists for this BSP. If it exists, then either the WKS_FILE
has been changed (presumably by the BSP), or the default path which includes
the MACHINE exists. In either case, we can assume the existence of such a file
indicates wic support is available for this BSP, so use it.

Signed-off-by: Christopher Larson <chris_larson@mentor.com>