systemd: serial-getty: fix no login prompt bug
This bug is observed randomly on first boots. This issues seems to occur
when udev does not generate an event to notify systemd that the serial
device is ready. And systemd keeps on waiting for the event and it times
out resulting in no serial login prompt. This is a workaround which
removed the dependency of systemd waiting on udev event.

Signed-off-by: Adnan Ali <Adnan_Ali@mentor.com>