sdk_auto_relocate: Use env setup script's location for SDK path
Use the location of environment-setup script to determine the path of
SDK/sysroots. This will allow us to use the script without it being auto-
relocated. Certain development environments use this script to extract
values of sysroot paths by just parsing the script and the script is
not sourced/executed.

http://jira.alm.mentorg.com:8080/browse/SB-7175

Signed-off-by: Fahad Usman <fahad_usman@mentor.com>