setup-ubuntu: install virtualenv for toaster setup
We want to use a virtualenv rather than installing into $PYTHONUSERBASE for
our toaster dependencies, so ensure that we have it available. This first
attempts to use apt-get, then falls back to installing virtualenv into
$PYTHONUSERBASE, as the python-virtualenv package is in 'universe' in the
package feeds, which won't be enabled on all systems, and I don't want to try
to script enabling it, even if we wanted to do so.

Signed-off-by: Christopher Larson <chris_larson@mentor.com>