gdbserver-gplv3: avoid reimplementing license whitelisting
oe-core has built in support for license whitelisting, so make use of it. We
still need our bits which remove gdbserver from the codebench-debug feature
when it's not buildable, however.

As a consequence of this, we now get messages from bitbake reminding us that it's being included despite the incompatible license:

    NOTE: INCLUDING gdb as buildable despite INCOMPATIBLE_LICENSE because it has been whitelisted
    NOTE: INCLUDING gdbserver-external as buildable despite INCOMPATIBLE_LICENSE because it has been whitelisted

I think this behavioral change is an improvement, since it keeps the user
aware of the exception to the rule.

This is needed to fix ALLOW_GPLV3_GDBSERVER, as by default we use the
standalone 'gdb' recipe for gdbserver, not 'gdbserver-external', but the
implementation didn't allow 'gdb' to build.

Signed-off-by: Christopher Larson <chris_larson@mentor.com>