mel: set shallow depth to zero if revs is set
This is only the default -- it could of course be overridden if one desires to
remove both the history of the specified revs and the revs selected by
branch+srcrev/depth.

With this, we no longer need to set depth manually for BSP-specific kernels
where we set revs.

JIRA: SB-8530

Signed-off-by: Christopher Larson <chris_larson@mentor.com>