toaster: use layer names from layer.conf
The current mechanism relies on folder names for
identification of a particular layer. This is not
suitable as folder names can be same over different
layers and tends to pollute the toaster setup as
it gets conflicting entries.
We change the behavior to make use of the real layer
name (BBFILE_COLLECTIONS) from each layer's layer.conf
which are bound to be unique for a correct bitbake
environment setup.

JIRA Ticket: SB-8582

Signed-off-by: Awais Belal <awais_belal@mentor.com>